--- openapi/constants.js
@@ -1,29 +1,0 @@
-const propertyErrata = [
-  { path: 'FactorDevice.links', skip: true, skipReason: 'Not currently supported' },
-  { path: 'Link.hints', skip: true, skipReason: 'Not currently supported' },
-  { path: 'User._links', skip: true, skipReason: 'Not currently supported' },
-  { path: 'UserGroup._embedded', skip: true, skipReason: 'Not currently supported' },
-  { path: 'UserGroup._links', skip: true, skipReason: 'Not currently supported' },
-  { path: 'UserGroupStats._links', skip: true, skipReason: 'Not currently supported' },
-  
-  { path: 'ActivationToken.activationToken', rename: 'token', renameReason: '.NET type name and member name cannot be identical' },
-  { path: 'TempPassword.tempPassword', rename: 'password', renameReason: '.NET type name and member name cannot be identical' },
-
-  { path: 'CallFactor.profile', hidesBaseMember: true },
-  { path: 'EmailFactor.profile', hidesBaseMember: true },
-  { path: 'HardwareFactor.profile', hidesBaseMember: true },
-  { path: 'PushFactor.profile', hidesBaseMember: true },
-  { path: 'SecurityQuestionFactor.profile', hidesBaseMember: true },
-  { path: 'SmsFactor.profile', hidesBaseMember: true },
-  { path: 'TokenFactor.profile', hidesBaseMember: true },
-  { path: 'TotpFactor.profile', hidesBaseMember: true },
-  { path: 'WebFactor.profile', hidesBaseMember: true },
-];
-
-const operationSkipList = [
-  { id: 'forgotPassword', reason: 'Revisit in alpha2 (#62)'},
-  { id: 'addRoleToUser', reason: 'Revisit in alpha2 (#102)'},
-];
-
-module.exports.propertyErrata = propertyErrata;
-module.exports.operationSkipList = operationSkipList;
