+++ openapi/errata.js
@@ -1,25 +1,64 @@
-const {
-  propertyErrata
-} = require('./constants');
+const propertyErrata = [
+  { path: 'FactorDevice.links', skip: true, skipReason: 'Not currently supported' },
+  { path: 'Link.hints', skip: true, skipReason: 'Not currently supported' },
+  { path: 'User._links', skip: true, skipReason: 'Not currently supported' },
+  { path: 'UserGroup._embedded', skip: true, skipReason: 'Not currently supported' },
+  { path: 'UserGroup._links', skip: true, skipReason: 'Not currently supported' },
+  { path: 'UserGroupStats._links', skip: true, skipReason: 'Not currently supported' },
+  
+  { path: 'ActivationToken.activationToken', rename: 'token', renameReason: '.NET type name and member name cannot be identical' },
+  { path: 'TempPassword.tempPassword', rename: 'password', renameReason: '.NET type name and member name cannot be identical' },
 
+  { path: 'CallFactor.profile', hidesBaseMember: true },
+  { path: 'EmailFactor.profile', hidesBaseMember: true },
+  { path: 'HardwareFactor.profile', hidesBaseMember: true },
+  { path: 'PushFactor.profile', hidesBaseMember: true },
+  { path: 'SecurityQuestionFactor.profile', hidesBaseMember: true },
+  { path: 'SmsFactor.profile', hidesBaseMember: true },
+  { path: 'TokenFactor.profile', hidesBaseMember: true },
+  { path: 'TotpFactor.profile', hidesBaseMember: true },
+  { path: 'WebFactor.profile', hidesBaseMember: true },
+];
 
-function shouldSkipProperty(property, infoLogger) {
+function getRenamedProperty(fullPath) {
+  let errata = propertyErrata.find(x => x.path === fullPath);
+  if (!errata) return;
+  if (!errata.rename) return;
+
+  return {
+    displayName: errata.rename,
+    reason: errata.renameReason
+  }
+}
+
+function shouldHideBaseMember(fullPath) {
+  let errata = propertyErrata.find(x => x.path === fullPath);
+  if (!errata) return;
+  if (!errata.hidesBaseMember) return;
+
+  return true;
+}
+
+function shouldSkipProperty(property) {
   if (property.model && property.model === 'object') {
-    infoLogger('Skipping property', property.fullPath, '(Reason: object properties are not supported)');
-    return true;
+    return {
+      reason: 'object properties are not supported'
+    };
   }
 
   if (typeof property.commonType === 'undefined') {
-    infoLogger('Skipping property', property.fullPath, '(Reason: properties without commonType are not supported)');
-    return true;
+    return {
+      reason: 'properties without commonType are not supported'
+    };
   }
 
   let propertyDetails = propertyErrata.find(x => x.path == property.fullPath);
-  if (!propertyDetails) return false;
-  if (!propertyDetails.skip) return false;
+  if (!propertyDetails) return null;
+  if (!propertyDetails.skip) return null;
 
-  infoLogger('Skipping property', property.fullPath, `(Reason: ${propertyDetails.skipReason})`);
-  return true;
+  return {
+    reason: propertyDetails.skipReason
+  };
 }
 
 const modelMethodSkipList = [
@@ -38,13 +77,31 @@ const modelMethodSkipList = [
   { path: 'Group.listUsers', reason: 'Implemented as IGroup.Users' },
 ];
 
-function shouldSkipMethod(method, infoLogger) {
-  let skipRule = modelMethodSkipList.find(x => x.path === method.fullPath);
-  if (!skipRule) return false;
+function shouldSkipModelMethod(fullPath) {
+  let skipRule = modelMethodSkipList.find(x => x.path === fullPath);
+  if (!skipRule) return null;
 
-  infoLogger('Skipping model method', method.fullPath, `(Reason: ${skipRule.reason})`);
-  return true;
+  return {
+    reason: skipRule.reason
+  };
+}
+
+const operationSkipList = [
+  { id: 'forgotPassword', reason: 'Revisit in alpha2 (#62)'},
+  { id: 'addRoleToUser', reason: 'Revisit in alpha2 (#102)'},
+];
+
+function shouldSkipOperation(operationId) {
+  let skipRule = operationSkipList.find(x => x.id === operationId);
+  if (!skipRule) return null;
+
+  return {
+    reason: skipRule.reason
+  };
 }
 
 module.exports.shouldSkipProperty = shouldSkipProperty;
-module.exports.shouldSkipMethod = shouldSkipMethod;
+module.exports.shouldSkipModelMethod = shouldSkipModelMethod;
+module.exports.getRenamedProperty = getRenamedProperty;
+module.exports.shouldHideBaseMember = shouldHideBaseMember;
+module.exports.shouldSkipOperation = shouldSkipOperation;
